fix enum display name and break down table
